Styles for molecules
