+++ src/pages/NewYear/molecules/NewYearCategoryMolecule/NewYearCategoryMolecule.module.css
@@ -1,0 +1,3 @@
+.newYear__list {
+  color: var(--text-color);
+}
