+++ src/pages/NewYear/molecules/NewYearHeaderMolecule/NewYearHeaderMolecule.module.css
@@ -1,0 +1,4 @@
+.newYear__buttons {
+  display: flex;
+  justify-content: space-evenly;
+}
